requirements.txt
# Draught Backend - Django & WebSocket Dependencies
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.0,<5.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg[binary]>=3.1,<4.0
django-environ>=0.11,<1.0
gunicorn>=21.0,<23.0
redis>=5.0,<6.0
drf-spectacular>=0.27,<0.28
